requirements.txt
asgiref==3.3.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.1.3
djangorestframework==3.12.2
gunicorn==20.1.0
idna==3.3
python-dotenv==0.20.0
pytz==2020.4
requests==2.28.1
sqlparse==0.4.1
urllib3==1.26.11
